+++ include/exadg/compressible_navier_stokes/driver.h
@@ -31,7 +31,7 @@
 #include <exadg/compressible_navier_stokes/user_interface/input_parameters.h>
 #include <exadg/functions_and_boundary_conditions/verify_boundary_conditions.h>
 #include <exadg/grid/mapping_degree.h>
-#include <exadg/grid/mapping_finite_element.h>
+#include <exadg/grid/mapping_dof_vector.h>
 #include <exadg/matrix_free/matrix_free_wrapper.h>
 #include <exadg/utilities/print_general_infos.h>
